+++ src/discord/client.rs
@@ -7,6 +7,7 @@ pub async fn build_client(
     handler: crate::discord::events::Handler,
 ) -> serenity::Result<Client> {
     let intents = GatewayIntents::DIRECT_MESSAGES
+        | GatewayIntents::DIRECT_MESSAGE_REACTIONS
         | GatewayIntents::GUILDS
         | GatewayIntents::GUILD_MESSAGES
         | GatewayIntents::MESSAGE_CONTENT